Created validation module as a middleware
