+++ BackEnd/db/product.db.js
@@ -33,23 +33,6 @@ const getProductByIdDb = async (id) => {
 	return collection.findOne({ _id: ObjectId(id) })
 }
 
-const validateDb = async (product) => {
-	const db = mdb.get().db(process.env.DB_NAME)
-	const collection = db.collection('product')
-
-	console.log('\nValidating...\n')
-
-	// * Searches the database using the name of the product to be inserted
-	return await collection.findOne({ product_name: product.product_name })
-}
-
-// * Searches the database using the name of the product to be updated
-const validateIdDb = async (id) => {
-	const db = mdb.get().db(process.env.DB_NAME)
-	const collection = db.collection('product')
-
-	return await collection.findOne({ _id: ObjectId(id) })
-}
 
 const updateProductDb = async (id, changes) => {
 	const db = mdb.get().db(process.env.DB_NAME)
@@ -71,6 +54,5 @@ module.exports = {
 	getProductByIdDb,
 	updateProductDb,
 	deleteProductDb,
-	validateDb,
-	validateIdDb,
+
 }
