+++ BackEnd/middleware/index.js
@@ -1,0 +1,7 @@
+const authMiddleware = require('./validate.middleware')
+const validateMiddleware = require('./validate.middleware')
+
+module.exports = {
+    authMiddleware,
+    validateMiddleware
+}
